Moves the formatting dates function into a shared TS file.
