+++ src/app/search/search.component.ts
@@ -3,7 +3,7 @@ import { Component, OnInit, OnDestroy } from "@angular/core";
 import { SearchService } from "../search.service";
 import { Subscription } from "rxjs";
 import { NgbModal } from "@ng-bootstrap/ng-bootstrap";
-import { DatePipe } from "@angular/common";
+import { Shared } from "../shared";
 import { environment } from "../../environments/environment";
 
 @Component({
@@ -23,12 +23,7 @@ export class SearchComponent implements OnInit, OnDestroy {
   nonProjectItem: any;
   sbURL = environment.sbmainURL;
 
-  // Define date formats
-  MONTH_YEAR_FORMAT = "MM/yyyy";
-  FULL_DATE_FORMAT = "MM/dd/yyyy";
-
-  // Create an instance of DatePipe
-  datePipe = new DatePipe("en-US");
+  shared: Shared;
 
   constructor(
     private searchService: SearchService,
@@ -49,41 +44,23 @@ export class SearchComponent implements OnInit, OnDestroy {
     return false;
   }
 
-  formatDate(value) {
-    try {
-      const segments = value.split("-");
-      switch (segments.length) {
-        case 2:
-          return (
-            this.datePipe.transform(`${value}-01`, this.MONTH_YEAR_FORMAT) ||
-            value
-          );
-        case 3:
-          return this.datePipe.transform(value, this.FULL_DATE_FORMAT) || value;
-        case 1:
-        default:
-          // Return the year if it is the only value
-          return value;
-      }
-    } catch (error) {
-      console.error(`Could not parse value: ${value}. Error: ${error}`);
-      return value;
-    }
-  }
-
   ngOnInit() {
     this.filteredResultsSubscription =
       this.searchService.filteredResults$.subscribe((filteredResults) => {
         this.results = filteredResults;
         for (const result of this.results) {
           if (result.dates.start_date) {
-            result.dates.start_date = this.formatDate(result.dates.start_date);
+            result.dates.start_date = this.shared.formatDate(
+              result.dates.start_date,
+            );
           }
           if (result.dates.end_date) {
-            result.dates.end_date = this.formatDate(result.dates.end_date);
+            result.dates.end_date = this.shared.formatDate(
+              result.dates.end_date,
+            );
           }
           if (result.dates.publication_date) {
-            result.dates.publication_date = this.formatDate(
+            result.dates.publication_date = this.shared.formatDate(
               result.dates.publication_date,
             );
           }
